Big refactor in folder and files structure Files are now correctly placed in a ModName folder UI file got split to better understanding each part Title of each file now have the GitHub link
